A number of changes to support the second FPGA project

+++ TrackletAlgorithm/MatchProcessor.h
@@ -875,7 +875,7 @@ void readTable_rDSS(ap_uint<width> table[depth]){
 // MatchCalculator
 template<TF::layerDisk Layer, TF::phiRegion PHI, TF::seed Seed> constexpr bool FMMask();
 template<TF::layerDisk Layer, TF::phiRegion PHI> constexpr uint32_t FMMask();
-template<TF::layerDisk Layer, TF::phiRegion PHI> constexpr uint32_t NPage();
+template<TF::layerDisk Layer, TF::phiRegion PHI> constexpr uint64_t NPage();
 template<TF::layerDisk Layer, TF::phiRegion PHI> constexpr uint32_t NPageSum();
 #include "MatchProcessor_parameters.h"
 
@@ -1200,7 +1200,7 @@ void MatchProcessor(BXType bx,
   ap_uint<2> iPage[nMEM];
   ap_uint<5> iMem[nMEM];
 
-  constexpr unsigned int npages = NPage<LAYER, PHISEC>();
+  constexpr uint64_t npages = NPage<LAYER, PHISEC>();
   for (unsigned int imem = 0; imem < nINMEM; imem++) {
     unsigned int nPages = (npages >> (3*imem))&7;
     for (unsigned int j = 0 ; j < nPages; j++){
